backend/requirements.txt
asyncio
python-socketio[asgi]
ollama
langchain
langchain-community
langchain-core
redis
eventlet
uvicorn